Clean up and simplify mutt/offlineimap
For now, I am sticking with using actual filesystem folders in mutt
instead of virtual folders with notmuch. It's much simpler, and for the
time being I don't have any reason to justify the extra complexity of
the virtual folders.

Notmuch is still integrated to perform searches. In the future if I ever
have a bajillion emails and using real folders is too slow or something,
I may switch back. The config file is still in place and only needs to
be uncommented in muttrc.