Update vimwiki configuration
Add default template and stylesheet. For now these are both just copies
of the defaults from vimwiki itself, but eventually they can be
customized and modified.

Also remove the read_only.vim plugin. It's kind of pointless and
annoying, in practice (in the case where 'readonly' is turned off, the
mappings still stick around).