Disable Jedi "smart" auto mappings
By default Jedi automatically inserts the "import" in "from <module>
import <item>" statements, which was super annoying, so I disabled it.

Also temporarily disabled Python auto-format on write, because the
condition 'executable("yapf")' is returning true even when yapf is not
installed in the current virtualenv. Need to find a more sophisticated
solution that is robust to the current virtual environment.