Split up projectionist config into multiple files
Now there is one file per "project type" for projectionist. This was
necessary as the projectionist_heuristics variable continued to grow.

This commit also improves support for C++ and CMake projects by
automatically detecting projects with a compile_commands.json file and
setting the 'path' variable based on the contents of that file.