Refactor python autoload file
Create an async autoload function that wraps the jobs API of neovim and
vim, and use that new function to define the python#set_path() function.
Since that function runs asynchronously (and therefore has negligible
performance impact), run it on every new Python buffer when possible.

When the jobs API is not available, run the set_path() function
synchronously but cache the result for each directory.