Revert compile_commands projectionist changes
Revert back to simply parsing the file for ALL include paths (ignoring
which file they correspond to). Doing it by file had some unintended
consequences, not least of which is the fact that compile_commands
doesn't apply to header (.h) files. So instead parse the whole file for
paths and apply them to any c or cpp filetype file that is opened.