Make gmail use native IMAP instead of offlineimap
My gmail account has gigabytes worth of emails, which is a lot to
download and sync. Since I don't particularly care about having offline
access to that account, it makes more sense to just use IMAP directly.
This is easy to change in the future if I change my mind (just modify a
few lines in the offlineimap config and the config.gmail file under the
mutt config directory).

This commit also includes a bunch of updates to mutt, including adding
the sidebar and adding more options. Many of the settings were taken
from Steve Losh's [The Homely Mutt].

[The Homely Mutt]: http://stevelosh.com/blog/2012/10/the-homely-mutt/#why-local-email