Update bash configuration
Add a local copy of the git-prompt.sh completion script and keep it in
the user's bash_completion directory so it is automatically loaded (as
long as the bash-completion version is 2.8 or later).

Also update .bashrc to be more compatible with macOS. This now sources
the Homebrew installed version of bash-completion and properly enables
colored `ls` output.