Use path on commandline to search with C-T
If a path or partial path is already present on the commandline when C-T
is pressed, start searching from within that path.

This is straightforward in most cases except for with `find`, which
requires the path to search in to be its first argument. To deal with
this, introduce the `%` placeholder character in the command strings
which gets replaced by the current path token.