Don't use recursive search when grepping notes
The -r flag will search all hidden subdirectories as well (including,
for example, Syncthing's metadata folders). Using a glob will ensure
that only top-level files are grepped.

Note that this will only work if there are not subdirectories (which
there shouldn't be). If that ever changes, though, this will need to be
revisited.