Use notmuch-mutt script for notmuch queries
This works in both mutt and neomutt and so is more portable, the only
downside is that it requires the notmuch-mutt script to be installed.
This is fairly easy though: it already exists as a package in Debian and
I currently have a PR[1] open with Homebrew to add the notmuch-mutt
package.

[1]: https://github.com/Homebrew/homebrew-core/pull/51778