Revert back to using shell function for git wrapper
Using a separate shell script as a wrapper seemed way too fragile since
it depended on modifying the PATH to a hardcoded value. Using a shell
function is definitely a better approach, but the problem was that I
couldn't get it to work with fish's shell completion.

After plenty of trial and error, I was finally able to get it to work!
So now I can have the best of both worlds.