[vim] Fix bug when finding zet completions in first column
When `col('.')` is `1` then we actually end up checking `a:line[-1]`
(i.e. the *end* of the line). This isn't what we want -- if we're in the
first column then we know completion isn't valid since both tag and link
completion require some kind of prefix before it. So we should only be
evaluating the contents of the line when `start` is non-zero.