vim: create C/C++ tags per file
I think this will be a much better solution because it only creates tags
for header files used in the actual source file. This means we have to
re-create the tags files often as headers are added/removed, but since
it's done asynchronously I don't think it's a big deal. This is better
than the prior solution which generated a single enormous tags file for
all of the system headers -- this made looking for tags slow and made
tag completion unusable.