mutt: don't display authentication-results header
According RFC8601 Section 4.1, the Authentication-Results should
typically *not* be displayed in a mail client unless it can be verified
that the header is legitimate. Because this header can be forged and not
all MTAs adhere to the specification and remove forged header fields, it
should only be trusted if it can be ensured that the receiving MTA
correctly adds this header and also correctly strips pre-existing
headers using the MTA's "Verifier" id.

Gandi (my mail provider at the time of this commit) does not add this
header, so there is no point in displaying it since any such header
cannot be trusted.