vim: update editorconfig plugin
Update the filepath matching logic to match against the fullpath. If a
glob pattern doesn't contain a / character then only match against the
file name (rather than the full path). If it does contain a / character,
then consider the glob relative to the directory containing the config
file being parsed.