nvim: replace signify with gitsigns
Gitsigns is implemented entirely in Lua which is built into Neovim and
because of this requires no external dependencies. By default, Gitsigns
also uses Neovim's built-in diff library and runs in-process via
LuaJIT's FFI module. This is unlike other similar plugins that need to
run `git-diff` as an external process which is less efficient, has
tighter bottlenecks and requires file IO.