nvim: lint fennel files in-process
This can't be done well using nvim-lint, since that plugin spawns new
processes to lint the files. All of my Fennel configuration files have a
hidden preamble and there is no way (presently) to do this strictly on
the command line. So instead, lint fennel files using fennel directly
in-process which takes the same parameters as moonwalk does.