nvim: update autocmd macro
The flags are now inferred to be strings at the beginning of the body
form, which more closely mimics the :autocmd command in Vim. The benefit
of this is that autocommands which don't use any flags (the majority of
them) no longer have to include an empty table as the fourth argument.

Before:

   (autocmd :group :event :pattern [] (body))

After:

   (autocmd :group :event :pattern (body))