nvim: update autocmd macro
The group name in the autocmd is now a symbol rather than a string. This
makes it possible to determine whether or not a group is given, which
means I can now use group-less autocommands (should I need to). Also
introduce a convention to append '#' to the names of autocommand groups
that might already exist (e.g. fugitive#). This mimics how the Fennel
compiler uses the # character for its auto gensym to avoid name
collisions.