fish: more fixes to prompt jobs
Use the handy builtin fish_job_summary function to detect when a job
ends. This avoids the race condition of first checking if a job exists,
then defining a function that fires when the job ends. During that
(small) interval, the job may have already ended and fish would show an
error message because the job does not exist.