fish: always update jobs after command runs
This solves for the case where a long running program (like Vim) is
backgrounded, then put back into the foreground, then closed. The jobs
indicator in the prompt should change in that case, but the output of
`jobs -l` is empty because no new job has been put into the background.