nvim: add exec macro that does not return a value
vim.cmd returns an empty string, which counts as a truthy value and thus
deletes an autocommand if it is used as the final form in an autocommand
callback.

Instead, use a exec macro that calls nvim_exec2 (the same thing that
vim.cmd does) without returning any value.

vim.cmd is still used for passing a table argument. It returns a value
in this case as well, but that is documented and expected.