jj: include diff in description buffer
The new draft_commit_description option in jj allows me to add the
scissor line and a diff in the description buffer. jj doesn't (yet) have
a way to automatically remove lines that aren't prefixed with the JJ:
comment marker so for now I use a ftplugin in Nvim to remove all lines
after the scissor line manually.